feat(seguimientos): horas decimales + diálogos rediseñados + badges dark mode
- Backend: horas_realizadas cambia de INTEGER a NUMERIC(5,1) (Flyway V18)
  para soportar medias horas (ej. 7.5h); validación @DecimalMin(0.5)/@DecimalMax(24)
- Toda la cadena Flutter actualizada: model, service, providers, screens, widgets
  (int → double) sin romper la UI existente
- Diálogo "Registrar parte": stepper ±0.5h, date picker inline, textarea con
  contador — reemplaza la pantalla aparte que quedaba descontextualizada
- Diálogo "Reportar incidencia": showDialog centrado con chips de tipo animados
  en lugar del bottom sheet sin aire
- Badges de estado en modo oscuro: colores sólidos precalibrados en lugar de
  withOpacity sobre fondo oscuro (que resultaba casi invisible)
- Fix chat: .tipo("TEXTO") añadido al builder Lombok de Mensaje para no violar
  la constraint NOT NULL de la columna tipo

+++ backend/tfg-nexus-api/src/main/java/com/tfg/api/models/dto/SeguimientoResponse.java
@@ -11,7 +11,7 @@ public record SeguimientoResponse(
     Long id,
     Long practicaId,
     LocalDate fechaRegistro,
-    Integer horasRealizadas,
+    Double horasRealizadas,
     String descripcion,
     String estado,
     Long validadoPorId,
